move each topology to the appropriate directory
