+++ tests/test_expert_system.py
@@ -10,8 +10,8 @@ from degann.expert.tags import (
     RequiredModelPrecision,
 )
 from degann.networks import IModel
-from degann.networks.topology.topology_parameters import TensorflowDenseNetParams
-from degann.networks.topology.compile_parameters import SingleNetworkCompileParams
+from degann.networks.topology.base_topology_configs import TensorflowDenseNetParams
+from degann.networks.topology.base_compile_configs import SingleNetworkCompileParams
 
 
 @pytest.fixture
