+++ tests/test_simple_equation.py
@@ -1,7 +1,7 @@
 import pytest
 import numpy as np
 from degann.networks.imodel import IModel
-from degann.networks.topology.topology_parameters import TensorflowDenseNetParams
+from degann.networks.topology.base_topology_configs import TensorflowDenseNetParams
 from tests.utils import array_compare, init_params
 from degann.equations import simple_equation
